+++ WPF_Library/Models/Products/Drinks/Juice/PulpyGrape.cs
@@ -10,7 +10,7 @@ namespace Library.Models.Products.Drinks.Juice
 {
     public class PulpyGrape : Drink, IJuice
     {
-        public PulpyGrape(int id, string name, float cost, float volume) : base(id, name, cost, volume)
+        public PulpyGrape(int id, string name, decimal cost, double volume) : base(id, name, cost, volume)
         {
         }
     }
