+++ WPF_Library/Models/Products/Drinks/Juice/PulpyMango.cs
@@ -10,7 +10,7 @@ namespace Library.Models.Products.Drinks.Juice
 {
     public class PulpyMango : Drink, IJuice
     {
-        public PulpyMango(int id, string name, float cost, float volume) : base(id, name, cost, volume)
+        public PulpyMango(int id, string name, decimal cost, double volume) : base(id, name, cost, volume)
         {
         }
     }
