+++ WPF_Library/Models/Products/Drinks/Juice/Sprite.cs
@@ -10,7 +10,7 @@ namespace Library.Models.Products.Drinks.Juice
 {
     public class Sprite : Drink, IJuice
     {
-        public Sprite(int id, string name, float cost, float volume) : base(id, name, cost, volume)
+        public Sprite(int id, string name, decimal cost, double volume) : base(id, name, cost, volume)
         {
         }
     }
