+++ WPF_Library/Models/Products/Drinks/Water/Water.cs
@@ -10,7 +10,7 @@ namespace Library.Models.Products.Drinks.Water
 {
     public class Water : Drink, IWater
     {
-        public Water(int id, string name, float cost, float volume) : base(id, name, cost, volume)
+        public Water(int id, string name, decimal cost, double volume) : base(id, name, cost, volume)
         {
         }
     }
