+++ WPF_Library/Models/Products/Snaks/Cold/DoritosNachoWithCheese.cs
@@ -10,7 +10,7 @@ namespace Library.Models.Products.Snacks.Cold
 {
     public class DoritosNachoWithCheese : Snack, IHot
     {
-        public DoritosNachoWithCheese(int id, string name, float cost, float weigth) : base(id, name, cost, weigth)
+        public DoritosNachoWithCheese(int id, string name, decimal cost, double weigth) : base(id, name, cost, weigth)
         {
         }
     }
