+++ WPF_Library/Models/Products/Snaks/Sweets/CaramelCake.cs
@@ -10,7 +10,7 @@ namespace Library.Models.Products.Snacks.Sweet
 {
     public class CaramelCake : Snack, ISweet
     {
-        public CaramelCake(int id, string name, float cost, float weigth) : base(id, name, cost, weigth)
+        public CaramelCake(int id, string name, decimal cost, double weigth) : base(id, name, cost, weigth)
         {
         }
     }
