+++ WPF_Library/Models/Products/Snaks/Sweets/CaramelCroissant.cs
@@ -10,7 +10,7 @@ namespace Library.Models.Products.Snacks.Sweet
 {
     public class CaramelCroissant : Snack, ISweet
     {
-        public CaramelCroissant(int id, string name, float cost, float weigth) : base(id, name, cost, weigth)
+        public CaramelCroissant(int id, string name, decimal cost, double weigth) : base(id, name, cost, weigth)
         {
         }
     }
