+++ WPF_Library/Models/Products/Snaks/Sweets/VanillaCake.cs
@@ -10,7 +10,7 @@ namespace Library.Models.Products.Snacks.Sweet
 {
     public class VanillaCake : Snack, ISweet
     {
-        public VanillaCake(int id, string name, float cost, float weigth) : base(id, name, cost, weigth)
+        public VanillaCake(int id, string name, decimal cost, double weigth) : base(id, name, cost, weigth)
         {
         }
     }
